+++ bin/deploy-github-page.js
@@ -3,6 +3,7 @@ const path = require('path');
 
 const ENV_FILE = '.env.local';
 const BASH_FILE = path.resolve(__dirname, './deploy-github-page.sh');
+const BASH_FILE_INCREMENT = path.resolve(__dirname, './deploy-github-page-increment.sh');
 
 
 function deployGithubPage(options) {
@@ -16,6 +17,8 @@ function deployGithubPage(options) {
   const token = options.token || env.DEPLOY_GITHUB_PAGE_TOKEN;
   const branch = options.branch || env.DEPLOY_GITHUB_PAGE_BRANCH;
   const commitMessage = options.message || env.DEPLOY_GITHUB_PAGE_COMMIT_MESSAGE || '';
+
+  const isIncrement = (options.increment || env.DEPLOY_GITHUB_PAGE_COMMIT_INCREMENT) === '1';
 
   console.log('[DEPLOY]', {
     repoName,
@@ -35,6 +38,27 @@ function deployGithubPage(options) {
      || !branch
   ) {
     console.error('[DEPLOY ERROR] 信息不全');
+    return;
+  }
+
+  if (isIncrement) {
+    const playgroundDir = options.playgroundDir || env.DEPLOY_GITHUB_PAGE_PLAYGROUND_DIR || '';
+    const targetDirName = options.targetDirName || env.DEPLOY_GITHUB_PAGE_TARGE_DIR_NAME || '';
+
+    if (!playgroundDir) {
+      console.error('[DEPLOY ERROR] 请提供 playgroundDir');
+      return;
+    }
+    if (!targetDirName) {
+      console.error('[DEPLOY ERROR] 请提供 targetDirName');
+      return;
+    }
+
+    execSync(`sh ${BASH_FILE_INCREMENT} ${token} ${repoName} ${userName} ${email} ${targetDir} "${branch}" "${commitMessage}" ${playgroundDir} ${targetDirName}`, {
+      cwd: process.cwd(),
+      encoding: 'utf-8',
+      stdio: 'inherit',
+    });
     return;
   }