+++ src/disk/index.ts
@@ -1,0 +1,1 @@
+export { getAvailableDiskSize } from './disk';