+++ src/stylelint-rules/index.ts
@@ -1,1 +1,2 @@
+export { getStylelintLinebreaksRules } from './linebreaks';
 export { getStylelintVendorPrefixRules } from './vendor-prefix';