+++ src/stylelint-rules/linebreaks.ts
@@ -1,0 +1,12 @@
+export function getStylelintLinebreaksRules(value = 'unix') {
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
+  const os = require('os');
+  const platform = os.platform();
+
+  if (platform === 'linux') {
+    return {};
+  }
+  return {
+    linebreaks: value,
+  };
+}