+++ test/git/submodule.spec.ts
@@ -1,0 +1,8 @@
+import { getSubmodulePathList } from '../../src';
+
+
+describe('getSubmodulePathList', () => {
+  it('getSubmodulePathList', () => {
+    expect(getSubmodulePathList()).toMatchObject([]);
+  });
+});